Updating Drawer Menu (#231)
* Externalizing contexts from App.tsx

* Saving progress on navigation bar

* Continuing with navigation update

* Fixing routes for remaining areas of the app

* Cleaning up imports

* Updating all pages to use Sudokuru title

* Adding Profile and Statistics pages to menu

* Adding CloseMenu option to Drawer

* Updating navigation for end game modal

* Updating header and sidebar

* Adding testIDs and updating Cypress tests

* Updating theme colors

* Making Drawer menu as small as possible

* Updating label
