+++ app/Components/Statistics/StatisticsButton.tsx
@@ -1,25 +1,22 @@
-import { MaterialCommunityIcons } from "@expo/vector-icons";
-import { Button, useTheme } from "react-native-paper";
+import { IconButton, useTheme } from "react-native-paper";
 import React from "react";
 import { useNavigation } from "@react-navigation/native";
+import { PreferencesContext } from "../../Contexts/PreferencesContext";
 
 const StatisticsButton = () => {
   const navigation: any = useNavigation();
-  const theme = useTheme();
+  const { updateCurrentPage } = React.useContext(PreferencesContext);
 
   return (
-    <Button
-      mode="contained"
+    <IconButton
+      icon="chart-line"
       testID={"ViewStatisticsPageButton"}
-      style={{ marginRight: 5 }}
-      onPress={() => navigation.navigate("Statistics")}
-    >
-      <MaterialCommunityIcons
-        name="chart-line"
-        size={24}
-        color={theme.colors.onPrimaryContainer}
-      />
-    </Button>
+      size={20}
+      onPress={() => {
+        updateCurrentPage("Statistics");
+        navigation.navigate("Statistics");
+      }}
+    />
   );
 };
