+++ app/Components/Sudoku Board/EndGameModal.tsx
@@ -64,8 +64,8 @@ const EndGameModal = (props: any) => {
           <Text style={styles(reSize).statisticValue}>{props.difficulty}</Text>
         </View>
       </View>
-      <Button mode="contained" onPress={() => navigation.navigate("PlayHome")}>
-        Go Home
+      <Button mode="contained" onPress={() => navigation.navigate("Play")}>
+        Play New Game
       </Button>
     </View>
   );
