+++ app/Navigation/StackNavigators/ProfileStackNavigator.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import ProfilePage from "../../Pages/ProfilePage";
+
+const Stack = createStackNavigator();
+
+const ProfileStackNavigator = () => {
+  return (
+    <Stack.Navigator
+      screenOptions={{
+        title: "Sudokuru",
+        headerShown: false,
+      }}
+    >
+      <Stack.Screen name="ProfilePage" component={ProfilePage} />
+    </Stack.Navigator>
+  );
+};
+
+export default ProfileStackNavigator;
