+++ app/Pages/SudokuPage.tsx
@@ -65,7 +65,6 @@ const SudokuPage = ({ route, navigation }: any) => {
   return (
     <SafeAreaProvider>
       <SafeAreaView>
-        <Header page={"Sudoku"} />
         <View style={homeScreenStyles.home}>
           <View style={styles.statisticsTitle}>
             {/* The game now required the info about it to be rendered, which is given in generateGame() */}
